Making image_saver into a component.
