+++ image_view/src/image_saver.cpp
@@ -32,246 +32,13 @@
 *  POSSIBILITY OF SUCH DAMAGE.
 *********************************************************************/
 
-#include <opencv2/highgui/highgui.hpp>
-
 #include <rclcpp/rclcpp.hpp>
-#include <camera_calibration_parsers/parse.h>
-#include <cv_bridge/cv_bridge.h>
-#include <image_transport/image_transport.h>
-#include <std_srvs/srv/empty.hpp>
-#include <std_srvs/srv/trigger.hpp>
 
-#include <boost/format.hpp>
+#include <memory>
 
-#include <chrono>
+#include "image_view/image_saver_node.hpp"
 
-namespace image_view
-{
-
-using std::placeholders::_1;
-using std::placeholders::_2;
-using std::placeholders::_3;
-
-class ImageSaverNode
-  : public rclcpp::Node
-{
-  boost::format g_format;
-  bool save_all_image, save_image_service;
-  std::string encoding;
-  bool request_start_end;
-  bool is_first_image_;
-  bool has_camera_info_;
-  size_t count_;
-  rclcpp::Time start_time_;
-  rclcpp::Time end_time_;
-  image_transport::CameraSubscriber cam_sub_;
-  image_transport::Subscriber image_sub_;
-  rclcpp::Service<std_srvs::srv::Empty>::SharedPtr save_srv_;
-  rclcpp::Service<std_srvs::srv::Trigger>::SharedPtr start_srv_;
-  rclcpp::Service<std_srvs::srv::Trigger>::SharedPtr end_srv_;
-
-  bool saveImage(const sensor_msgs::msg::Image::ConstSharedPtr & image_msg, std::string & filename)
-  {
-    cv::Mat image;
-    try {
-      image = cv_bridge::toCvShare(image_msg, encoding)->image;
-    } catch(cv_bridge::Exception) {
-      RCLCPP_ERROR(
-        this->get_logger(), "Unable to convert %s image to %s",
-        image_msg->encoding.c_str(), encoding.c_str());
-      return false;
-    }
-
-    if (!image.empty()) {
-      try {
-        filename = (g_format).str();
-      } catch (...) {
-        g_format.clear();
-      }
-
-      try {
-        filename = (g_format % count_).str();
-      } catch (...) {
-        g_format.clear();
-      }
-
-      try { 
-        filename = (g_format % count_ % "jpg").str();
-      } catch (...) {
-        g_format.clear();
-      }
-
-      if (save_all_image || save_image_service ) {
-        cv::imwrite(filename, image);
-        RCLCPP_INFO(this->get_logger(), "Saved image %s", filename.c_str());
-
-        save_image_service = false;
-      } else {
-        return false;
-      }
-    } else {
-      RCLCPP_WARN(this->get_logger(), "Couldn't save image, no data!");
-      return false;
-    }
-
-    return true;
-  }
-
-public:
-  ImageSaverNode(const rclcpp::NodeOptions & options)
-    : rclcpp::Node("image_saver_node", options),
-      is_first_image_(true), has_camera_info_(false), count_(0)
-  {
-    auto topic = rclcpp::expand_topic_or_service_name("image", this->get_name(), this->get_namespace());
-
-    // Useful when CameraInfo is being published
-    cam_sub_ = image_transport::create_camera_subscription(
-      this, topic, std::bind(
-        &ImageSaverNode::callbackWithCameraInfo, this, std::placeholders::_1, std::placeholders::_2),
-      "raw");
-
-    // Useful when CameraInfo is not being published
-    image_sub_ = image_transport::create_subscription(
-      this, topic, std::bind(
-        &ImageSaverNode::callbackWithoutCameraInfo, this, std::placeholders::_1),
-      "raw");
-
-    std::string format_string;
-    format_string = this->declare_parameter("filename_format", std::string("left%04i.%s"));
-    encoding = this->declare_parameter("encoding", std::string("bgr8"));
-    save_all_image = this->declare_parameter("save_all_image", true);
-    request_start_end = this->declare_parameter("request_start_end", false);
-    g_format.parse(format_string);
-
-    if (request_start_end && !save_all_image) {
-      RCLCPP_WARN(
-        this->get_logger(), "'request_start_end' is true, so overwriting 'save_all_image' as true");
-    }
-
-    save_srv_ = this->create_service<std_srvs::srv::Empty>(
-      "save", std::bind(&ImageSaverNode::service, this, _1, _2, _3));
-
-    // FIXME(unkown): This does not make services appear
-    // if (request_start_end) {
-    start_srv_ = this->create_service<std_srvs::srv::Trigger>(
-        "start", std::bind(&ImageSaverNode::callbackStartSave, this, _1, _2, _3));
-    end_srv_ = this->create_service<std_srvs::srv::Trigger>(
-        "end", std::bind(&ImageSaverNode::callbackEndSave, this, _1, _2, _3));
-    // }
-  }
-
-  bool service(
-    const std::shared_ptr<rmw_request_id_t> request_header,
-    const std::shared_ptr<std_srvs::srv::Empty::Request> request,
-    std::shared_ptr<std_srvs::srv::Empty::Response> response)
-  {
-    (void)request_header;
-    (void)request;
-    (void)response;
-    save_image_service = true;
-    return true;
-  }
-
-  bool callbackStartSave(
-    const std::shared_ptr<rmw_request_id_t> request_header,
-    const std::shared_ptr<std_srvs::srv::Trigger::Request> request,
-    std::shared_ptr<std_srvs::srv::Trigger::Response> response)
-  {
-    (void)request_header;
-    (void)request;
-    RCLCPP_INFO(this->get_logger(), "Received start saving request");
-    start_time_ = this->now();
-    end_time_ = rclcpp::Time(0);
-
-    response->success = true;
-    return true;
-  }
-
-  bool callbackEndSave(
-    const std::shared_ptr<rmw_request_id_t> request_header,
-    const std::shared_ptr<std_srvs::srv::Trigger::Request> request,
-    std::shared_ptr<std_srvs::srv::Trigger::Response> response)
-  {
-    (void)request_header;
-    (void)request;
-    RCLCPP_INFO(this->get_logger(), "Received end saving request");
-    end_time_ = this->now();
-
-    response->success = true;
-    return true;
-  }
-
-  void callbackWithoutCameraInfo(const sensor_msgs::msg::Image::ConstSharedPtr & image_msg)
-  {
-    if (is_first_image_) {
-      is_first_image_ = false;
-
-      // Wait a tiny bit to see whether callbackWithCameraInfo is called
-      rclcpp::sleep_for(std::chrono::milliseconds(1));
-    }
-
-    if (has_camera_info_) {
-      return;
-    }
-
-    // saving flag priority:
-    //  1. request by service.
-    //  2. request by topic about start and end.
-    //  3. flag 'save_all_image'.
-    if (!save_image_service && request_start_end) {
-      if (start_time_ == rclcpp::Time(0)) {
-        return;
-      } else if (start_time_ > image_msg->header.stamp) {
-        return;  // wait for message which comes after start_time
-      } else if ((end_time_ != rclcpp::Time(0)) && (end_time_ < image_msg->header.stamp)) {
-        return;  // skip message which comes after end_time
-      }
-    }
-
-    // save the image
-    std::string filename;
-    if (!saveImage(image_msg, filename)) {
-      return;
-    }
-
-    count_++;
-  }
-
-  void callbackWithCameraInfo(
-    const sensor_msgs::msg::Image::ConstSharedPtr & image_msg,
-    const sensor_msgs::msg::CameraInfo::ConstSharedPtr & info)
-  {
-    has_camera_info_ = true;
-
-    if (!save_image_service && request_start_end) {
-      if (start_time_ == rclcpp::Time(0)) {
-        return;
-      } else if (start_time_ > image_msg->header.stamp) {
-        return;  // wait for message which comes after start_time
-      } else if ((end_time_ != rclcpp::Time(0)) && (end_time_ < image_msg->header.stamp)) {
-        return;  // skip message which comes after end_time
-      }
-    }
-
-    // save the image
-    std::string filename;
-    if (!saveImage(image_msg, filename)) {
-      return;
-    }
-
-    // save the CameraInfo
-    if (info) {
-      filename = filename.replace(filename.rfind("."), filename.length(), ".ini");
-      camera_calibration_parsers::writeCalibration(filename, "camera", *info);
-    }
-
-    count_++;
-  }
-};
-
-}  // namespace image_view
-
-int main(int argc, char** argv)
+int main(int argc, char ** argv)
 {
   using image_view::ImageSaverNode;
